+++ CHANGELOG.md
@@ -7,7 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [2.1.x] - 2020-04-04
+## [2.1.2]
+
+### Added
+
+- Added PHP7.4 typed properties support
+
+### Changed
+
+- Changed doctrine inflector by symfony inflector
+- Fixed `PhpParser\Comment` deprecations
+- Improved logs
+- Fixed nullable method documentation return type
+
+### Removed
+
+## [2.1.1] - 2020-04-04
 
 ### Added
 
@@ -18,7 +33,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Fixed nullable method documentation return type
 - Improved commands output
 - Improved self composer dependencies
 - Changed `RuntimeException` and `LogicException` to more accurate ones